Move the pane tree to Core behind IPaneView
The tree model (Pane, Branch, Split, Tree, PaneId) lived in App only
because a pane held the real TerminalControl, so its structural
rules — replace, remove with sibling promotion, zoom invalidation —
could never run in Tests.exe, and the split-action decisions had to
live away from the tree they describe. A pane now carries its
control split by role: an IInspectable handle that keeps it alive,
and an IPaneView — the small interface of what a pane's owners do
with it (surface access, detach, rehost, focus visuals, opaque
underlay, take focus) — which TerminalControl implements beside
ISurfaceView. That makes the tree a plain Core class, and most call
sites read better than before (p.view->Detach() for p.Impl()->
Detach()); the few that need TerminalControl-only methods go through
the one sanctioned downcast, ControlOf, replacing the Impl() whose
name said nothing. test_tree.cpp covers the walks and both
mutations with empty handles and null views.
