+++ App/Tabs/Panes/Pane.h
@@ -1,30 +1,26 @@
 #pragma once
 
+#include <Panes/Pane.h>
 #include "Tabs/Panes/PaneId.h"
 #include "Terminal/TerminalControl.xaml.h"
-#include <winrt/Microsoft.UI.Xaml.h>
 
 namespace winrt::GhosttyWin32::implementation {
 
-// A single terminal — the tree's leaf. One pane hosts exactly one
-// TerminalControl, which in turn owns exactly one ghostty surface.
-// That 1:1:1 chain is ghostty's own design premise (its split tree
-// is SplitTree<SurfaceView>), not a host simplification, so the
-// member is typed as the control rather than a generic UIElement.
-// Layout code (SplitPanel) still uses `control` as the UIElement it
-// is; logic code reaches the implementation through Impl() instead
-// of a try_as / get_self pair at every site.
-struct Pane {
-    winrt::GhosttyWin32::TerminalControl control{ nullptr };
-    PaneId id{};
+// The tree's leaf. Core owns the type (Core/Panes/Pane.h): a pane
+// carries its control as an IInspectable `handle` (ownership) plus
+// an IPaneView `view` (what the pane's owners do with it), so the
+// tree never names TerminalControl and lives in Core. In this app
+// both refer to the same TerminalControl; TabFactory sets them when
+// it makes the pane.
+using Pane = core::panes::Pane;
 
-    // The control's implementation object, or nullptr for an empty
-    // pane. Borrowed: the projected `control` keeps it alive.
-    implementation::TerminalControl* Impl() const noexcept {
-        return control
-            ? winrt::get_self<implementation::TerminalControl>(control)
-            : nullptr;
-    }
-};
+// The concrete control behind a pane, for the few callers that need
+// TerminalControl-only methods (IME notifications, composition
+// handle, the search box, InnerPanel). The one sanctioned downcast:
+// this app puts exactly one kind of view into panes, so `view` is
+// always a TerminalControl. Borrowed — `handle` keeps it alive.
+inline TerminalControl* ControlOf(Pane const& p) noexcept {
+    return p.view ? static_cast<TerminalControl*>(p.view) : nullptr;
+}
 
 }  // namespace winrt::GhosttyWin32::implementation
