+++ Tests/Tests.vcxproj
@@ -174,6 +174,7 @@
     <ClCompile Include="test_window_decorations.cpp" />
     <ClCompile Include="test_background_opacity.cpp" />
     <ClCompile Include="test_splits.cpp" />
+    <ClCompile Include="test_tree.cpp" />
     <ClCompile Include="test_tab_drag.cpp" />
     <ClCompile Include="test_pressed_tab.cpp" />
     <ClCompile Include="test_trigger_label.cpp" />
